% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{buylimit}
\alias{buylimit}
\title{Place a Buy Limit Order}
\usage{
buylimit(market, quantity, rate)
}
\description{
The \code{buylimit} function places a buy order onto the 
C-Cex crypto-currency exchange \url{https://c-cex.com}. This function
only works if you have set up authentication with the 
\code{\link{ccex_authenticate}} function.
}
\examples{
\dontrun{
# Buy one bitcoin for 1169 dollars.
order = buylimit("btc-usd", 1, 1169)
}
}
\seealso{
\code{link{ccex_authenticate}}
}
