% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{cancel}
\alias{cancel}
\title{Cancel an Open Order}
\usage{
cancel(uuid)
}
\arguments{
\item{uuid}{the uuid of the order you would like to cancel.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{always NULL}
}
\code{\link{getopenorders}} \code{\link{getorderhistory}} 
\code{\link{buylimit}} \code{\link{selllimit}}
}
\description{
The \code{cancel} function cancels an open order on the
C-Cex crypto-currency exchange \url{https://c-cex.com}. This function
is called after providing information to authenticate your account and 
after an order is placed using either 
the \code{link{buylimit}} or \code{link{selllimit}} functions.
}
\examples{
\dontrun{
cancel(uuid) 
}
}
\references{
\url{https://c-cex.com/?id=api#cancel}
}
\seealso{
\code{\link{ccex_authenticate}} \code{\link{getorder}}
}
