% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\docType{package}
\name{ccex-package}
\alias{ccex-package}
\alias{ccex}
\title{Client for the C-Cex Crypto-Currency Exchange}
\description{
This software is in no way affiliated, endorsed, or approved by
the C-Cex crypto-currency exchange or any of its affiliates. It comes with
absolutely no warranty and should not be used in actual trading
unless the user can read and understand the source and know what you are
doing.

Package 'ccex' is an R implementation of the REST interface used by the C-Cex
crypto-currency exchange \url{https://c-cex.com/}. It provides functions 
for all endpoints currently (as of May 16, 2017) supported by the 
exchange. This includes the ability 
to retrieve price, volume, and orderbook information as well as the ability
to trade crypto-currencies.

Calls to the exchange are categorized as either public, which includes 
requests for price, volume, and order book information, and private, which 
includes all requests requiring an account including placing buy or sell 
orders. Public calls can be used directly by installing the package. 
Private calls require creating an account at 
\url{https://c-cex.com/?id=reg} and creating an API and secret key with 
appropriate permissions.

Private calls retrieve the API and secret key using the CCEX_API_KEY and 
CCEX_SECRET_KEY environment variables. These may be set by the user before 
opening the R session or, they can be set using the 'ccex_authenticate' 
function.

Public Function Calls
\itemize{
\item{coinnames: }{coin names for all currencies}
\item{getballancedistribution: }{exchange's wallet balance distribution}
\item{getmarkethistory: }{recent history for a market}
\item{getmarkets: }{available markets and other meta-data}
\item{getmarketsummaries: }{summary of active markets}
\item{getorderbook: }{order book for a market}
\item{pairs: }{available trading pairs}
\item{prices: }{prices for all markets}
\item{volume: }{volume of trades for a currency}
}
Private Function Calls
\itemize{
\item{buylimit: }{place a buy limit order}
\item{cancel: }{cancel an open order}
\item{ccex_authenticate: }{provide user authentication data}
\item{getbalance: }{account balance for a specified currency}
\item{getbalances: }{account balances for currencies}
\item{getopenorders: }{order data for all open orders}
\item{getorderdata: }{order data for a specified order}
\item{getorderhistory: }{recent order history for an account }
\item{mytrades: }{detailed recent trading history for a market}
\item{selllimit: }{place a sell limit order}
}
}
\references{
\url{https://c-cex.com/}
\url{https://github.com/kaneplusplus/ccex}
}
