% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{coinnames}
\alias{coinnames}
\title{Retrieve Full Coin Names for All Currencies}
\usage{
coinnames()
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result: }{a \code{data.frame} with the ticker and corresponding full
                 name of the currency.}
}
}
\description{
The \code{coinnames} function returns the full coin names
for all available currencies on the C-Cex crypto-currency exchange
\url{https://c-cex.com}.
}
\examples{
\dontrun{
cns = coinnames()$result
head(cns)
}
}
\references{
\url{https://c-cex.com/?id=api#coinnames.json}
}
