% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{getbalance}
\alias{getbalance}
\title{Get the Balance a Specified Currency}
\usage{
getbalance(currency = "btc")
}
\arguments{
\item{currency}{the currency to get your account balance for.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} with the currency, balance, 
   available funds, the amount of any pending transactions, and a
   crypographic address that can be used to receive more funding.
 }
}
}
\description{
The \code{getbalance} function retrieves the account balance
for a specified currency on the C-Cex crypto-currency 
exchange \url{https://c-cex.com}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
getbalance("btc") 
}
}
\references{
\url{https://c-cex.com/?id=api#getbalance}
}
\seealso{
\code{\link{ccex_authenticate}}
}
