% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{getbalancedistribution}
\alias{getbalancedistribution}
\title{Exchange's Wallet Balance Distribution}
\usage{
getbalancedistribution(currency)
}
\arguments{
\item{currency}{the currency to get the balance distrubtion for.}
}
\description{
The \code{getbalancedistribution} function retrieves the
C-Cex crypto-currency exchange's wallet balance distribution for a specified
currency.
}
\examples{
\dontrun{
mh = getbalancedistribution("grc")$result
head(ms)
}
}
\references{
\url{https://c-cex.com/?id=api#getbalancedistribution}
}
