% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{getbalances}
\alias{getbalances}
\title{Get the Balances for All Currencies}
\usage{
getbalances()
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} with the currencies, balances, 
   available funds, the amount of any pending transactions, and 
   crypographic addresses that can be used to receive funding.
 }
}
}
\description{
The \code{getbalances} function retrieves the account balance
for all currencies on the C-Cex crypto-currency 
exchange \url{https://c-cex.com}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
balances = getbalances()$result
}
}
\references{
\url{https://c-cex.com/?id=api#getbalances}
}
\seealso{
\code{\link{ccex_authenticate}}
}
