% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{getopenorders}
\alias{getopenorders}
\title{Get Data for All Open Orders}
\usage{
getopenorders(market = NULL)
}
\arguments{
\item{market}{(optional) the market on which you would like to see all 
open orders. If not specified, then all open orders
for all markets are returned.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing information about the 
   open orders including (but not limited to) the market, quantity remaining
   in the order, the type of order, and when the order was opened.
 }
}
}
\description{
The \code{getopenorders} function retrieves all open orders
on the C-Cex crypto-currency 
exchange \url{https://c-cex.com}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
getopenorders("btc-usd")
}
}
\references{
\url{https://c-cex.com/?id=api#getorder}
}
\seealso{
\code{\link{ccex_authenticate}} \code{\link{getorder}}
\code{\link{mytrades}}
}
