% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{getorder}
\alias{getorder}
\title{Get Order Data}
\usage{
getorder(uuid)
}
\arguments{
\item{uuid}{the uuid of the order.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing information about the 
   open order including (but not limited to) the market, quantity remaining
   in the order, the type of order, and when the order was opened.
 }
}
}
\description{
The \code{getorder} function retrieves open order data 
on the C-Cex crypto-currency 
exchange \url{https://c-cex.com}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
getorder(uuid)
}
}
\references{
\url{https://c-cex.com/?id=api#getorder}
}
\seealso{
\code{\link{ccex_authenticate}} \code{\link{getopenorders}}.
}
