% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{getorderhistory}
\alias{getorderhistory}
\title{Recent Order History for an Account}
\usage{
getorderhistory(market = NULL, count = NULL)
}
\arguments{
\item{market}{(optional) the market on which you would like to see all 
open orders. If not specified, then completed orders for all markets are 
returned.}

\item{count}{(optional) the number of records to return.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing data about 
   previously completed orders including the order uuid, the exchange
   the time of the order, the order type, the limit, the quantity, the
   quantity remaining, the commission, the price, the price per unit,
   and whether or not it was a conditional trade.
 }
}
}
\description{
The \code{getorderhistory} function retrieves order history
data on the C-Cex crypto-currency exchange \url{https://c-cex.com}. This 
function can be used after you provide information for authentication.
}
\examples{
\dontrun{
getorderhistory()
}
}
\references{
\url{https://c-cex.com/?id=api#getorderhistory}
}
\seealso{
\code{\link{ccex_authenticate}}
}
