% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{mytrades}
\alias{mytrades}
\title{Detailed Trading History for a Market}
\usage{
mytrades(market, limit = 3)
}
\arguments{
\item{market}{the market on which you would like to see the history data.}

\item{limit}{(optional) the number of records to return.}
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a \code{data.frame} providing detailed information about
   trades on the market including (but not limited to) the order type,
   the time of the trade, the trade price, the quantity, and the uuid.
 }
}
}
\description{
The \code{mytrades} function retrieves a detailed
trading history for a specified market on the C-Cex crypto-currency 
exchange \url{https://c-cex.com}. This function
can be used after you provide information for authentication.
}
\examples{
\dontrun{
mytrades("btc-usd")
}
}
\seealso{
\code{\link{ccex_authenticate}} \code{\link{getorderhistory}}
}
