% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{pairs}
\alias{pairs}
\title{The List of Available Trading Pairs}
\usage{
pairs()
}
\value{
A named list with the following structure:
\itemize{
 \item{success: }{was the request successful?}
 \item{message: }{if the request was not successful, what was the error?}
 \item{result:  }{a character vector of the available markets.}
}
}
\description{
The \code{pairs} function returns all currency-pairs 
currently available on the C-Cex crypto-currency exchange
\url{https://c-cex.com}.
}
\details{
The \code{pairs} function returns a named list. The \code{result} 
element is a vector of names for all markets on the exchange. The
format for a market is "currency1-currency2" 
where currency2 is the reference currency. 

As an example, if you want to buy U.S. dollars with bitcoins you would 
create a "buy" order for the market denoted "usd-btc".
}
\examples{
markets = pairs()$result
}
\references{
\url{https://c-cex.com/t/prices.json}
}

