% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{prices}
\alias{prices}
\title{Retrieve All Trading Pairs Data}
\usage{
prices()
}
\description{
The \code{prices} function returns all trading pairs data
currently available on the C-Cex crypto-currency exchange.
}

