% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{pairs}
\alias{pairs}
\title{Available Trading Pairs}
\usage{
pairs()
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result:  }{a character vector of the available markets.}
}
}
\description{
The \code{pairs} function returns all currency-pairs 
currently available on the C-Cex crypto-currency exchange
\url{https://c-cex.com}.
}
\details{
The \code{pairs} function returns a named list. The \code{result} 
element is a vector of names for all markets on the exchange. The
format for a market is "currency1-currency2" 
where currency2 is the reference currency. 

As an example, if you want to buy U.S. dollars with bitcoins you would 
create a "buy" order for the market denoted "usd-btc".
}
\examples{
\dontrun{
markets = pairs()$result
print(markets)
}
}
\references{
\url{https://c-cex.com/?id=api#prices.json}
}
