% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccex.r
\name{prices}
\alias{prices}
\title{Retrieve All Trading Pairs Data}
\usage{
prices()
}
\value{
A named list with the following elements:
\itemize{
 \item{success: }{a boolean indicating if the request successful?}
 \item{message: }{a string describing the error if the request was not 
                  successful, otherwise and empty string."}
 \item{result: }{a \code{data.frame} with sell and buy data for each market.}
}
}
\description{
The \code{prices} function returns all trading pairs data
currently available on the C-Cex crypto-currency exchange
\url{https://c-cex.com}.
}
\examples{
\dontrun{
all_prices = prices()$result
head(all_prices)
}
}
\references{
\url{https://c-cex.com/?id=api#prices.json}
}
